enrichment_scripts/requirements-phantombuster.txt
# PhantomBuster MCP Integration Dependencies
# Install with: pip install -r requirements-phantombuster.txt

requests>=2.31.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0

